move MetadataView into metadata
