+++ src/metadata/src/lib.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// use crate::{DropSchemaError, DropStrategy};
 use binary::Binary;
 use chashmap::CHashMap;
 use kernel::{SystemError, SystemResult};
@@ -27,7 +26,23 @@ use std::{
         Arc, RwLock,
     },
 };
-use storage::{Database, InMemoryDatabase, InitStatus, PersistentDatabase};
+use storage::{Database, FullSchemaId, FullTableId, InMemoryDatabase, InitStatus, PersistentDatabase};
+
+pub trait MetadataView {
+    fn schema_exists<S: AsRef<str>>(&self, schema_name: &S) -> FullSchemaId;
+
+    fn table_exists<S: AsRef<str>, T: AsRef<str>>(&self, schema_name: &S, table_name: &T) -> FullTableId;
+
+    fn table_columns<I: AsRef<(Id, Id)>>(&self, table_id: &I) -> Result<Vec<ColumnDefinition>, ()>;
+
+    fn column_ids<I: AsRef<(Id, Id)>, N: AsRef<str> + PartialEq<N>>(
+        &self,
+        table_id: &I,
+        names: &[N],
+    ) -> Result<(Vec<Id>, Vec<String>), ()>;
+
+    fn column_defs<I: AsRef<(Id, Id)>>(&self, table_id: &I, ids: &[Id]) -> Vec<ColumnDefinition>;
+}
 
 const SYSTEM_CATALOG: &'_ str = "system";
 // CREATE SCHEMA DEFINITION_SCHEMA
