+++ src/query_executor/Cargo.toml
@@ -13,6 +13,7 @@ data_manager = { path = "../data_manager" }
 expr_eval = { path = "../expr_eval" }
 kernel = { path = "../kernel" }
 meta_def = { path = "../meta_def" }
+metadata = { path = "../metadata" }
 plan = { path = "../plan" }
 protocol = { path = "../protocol" }
 repr = { path = "../repr" }
