+++ src/query_executor/src/ddl/drop_schema.rs
@@ -12,12 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::sync::Arc;
-
 use data_manager::DataManager;
 use plan::SchemaId;
 use protocol::{results::QueryEvent, Sender};
 use sql_model::{DropSchemaError, DropStrategy};
+use std::sync::Arc;
 use storage::Database;
 
 pub(crate) struct DropSchemaCommand<D: Database> {
